remove profiling from nix build for now

+++ scripts/dev/test-nix.sh
@@ -20,7 +20,8 @@ TIMESTAMP=$(date '+%Y-%m-%d_%H:%M')
 LOGFILE="ortholang_${TEST_FILTER}_${TIMESTAMP}.log"
 
 nix-build $NIX_ARGS 2>&1 | tee -a $LOGFILE
-
+code0=$?
+[[ $code0 == 0 ]] || exit $code0
 
 bin-run() {
   rm -f ortholang.log
